Crate sc_sync_state_rpc

A RPC handler to create sync states for light clients.

Currently only usable with BABE + GRANDPA.

Usage

To use the light sync state, it needs to be added as an extension to the chain spec:

use sc_sync_state_rpc::LightSyncStateExtension;

#[derive(Default, Clone, serde::Serialize, serde::Deserialize, sc_chain_spec::ChainSpecExtension)]
#[serde(rename_all = "camelCase")]
pub struct Extensions {
   light_sync_state: LightSyncStateExtension,
}

type ChainSpec = sc_chain_spec::GenericChainSpec<(), Extensions>;

If the LightSyncStateExtension is not added as an extension to the chain spec, the SyncStateRpcHandler will fail at instantiation.

Modules

gen_client

The generated client module.

Structs

LightSyncState

Hardcoded infomation that allows light clients to sync quickly.

SyncStateRpcApiClient

The Client.

SyncStateRpcHandler

The handler for sync state RPC calls.

Enums

Error

Error type used by this crate.

Traits

SyncStateRpcApi

An api for sync state RPC calls.

Type Definitions

LightSyncStateExtension

The light sync state extension.