Trait sc_sync_state_rpc::SyncStateRpcApiServer

pub trait SyncStateRpcApiServer: Sized + Send + Sync + 'static {
    fn system_gen_sync_spec(&self, raw: bool) -> RpcResult<String>;

    fn into_rpc(self) -> RpcModule<Self> { ... }
}

Server trait implementation for the SyncStateRpcApi RPC API.

Required methods

fn system_gen_sync_spec(&self, raw: bool) -> RpcResult<String>

Returns the JSON serialized chainspec running the node, with a sync state.

Provided methods

fn into_rpc(self) -> RpcModule<Self>

Collects all the methods and subscriptions defined in the trait and adds them into a single RpcModule.